requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
python-nmap>=0.7.1
scapy>=2.5.0
eventlet>=0.35.1
